Make `test` the Test namespace root
Update the namespaces used in the file & locations to match.
